fix: 임시모임 컴파일 안 되게 설정 및 변경된 유저 패키지 적용 - build.gradle : 임시모임 기능 컴파일에서 제외
