--- src/main/java/GGUM_Team3/Server/domain/tag/hashtag/dto/HashtagMeetingsDTO.java
@@ -1,15 +1,0 @@
-package GGUM_Team3.Server.domain.tag.hashtag.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class HashtagMeetingsDTO {
-    private String hashtagName;
-    private List<SearchForMeetingWithHashtagsDTO> tempMeetings;
-}
